pyproject.toml
[project]
name = "adk-deepagents"
version = "0.1.0"
description = "Re-implementation of deepagents using Google ADK primitives"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0",
    "wcmatch>=8.5",
]

[project.optional-dependencies]
skills = [
    "adk-skills-agent>=0.1.0",
]
browser = [
    "google-adk[mcp]>=1.0.0",
]

[project.scripts]
adk-deepagents = "adk_deepagents.cli.main:cli_main"

[dependency-groups]
dev = [
    "adk-skills-agent>=0.1.0",
    "litellm>=1.81.13",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-timeout>=2.0",
    "ruff>=0.4",
    "ty>=0.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adk_deepagents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not llm and not browser'"
markers = [
    "integration: integration tests that test multiple components together",
    "llm: tests requiring a live LLM API key (deselect with '-m \"not llm\"')",
    "browser: tests requiring Playwright MCP + browser (npx @playwright/mcp, chromium)",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP007", # Use X | Y for union types (not always desired)
]

[tool.ruff.lint.isort]
known-first-party = ["adk_deepagents"]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
# Third-party library type stubs (google-adk) cause false positives
invalid-argument-type = "warn"
# Optional dependency (adk-skills-agent) may not be installed
unresolved-import = "warn"
# MCP toolset attributes resolved at runtime (incomplete stubs)
unresolved-attribute = "warn"
# Test files use assertions that the type checker can't narrow through
unsupported-operator = "warn"
not-subscriptable = "warn"
